Divided unit test in files
